fix(client): Rename list indices to list indexes
Closes #453
