+++ src/main/scala/algolia/definitions/DeleteDefinition.scala
@@ -103,52 +103,3 @@ case class DeleteByDefinition(
   override def options(requestOptions: RequestOptions): DeleteByDefinition =
     copy(requestOptions = Some(requestOptions))
 }
-
-trait DeleteDsl {
-
-  implicit val formats: Formats
-
-  case object delete {
-
-    //Index
-    def index(index: String): DeleteIndexDefinition =
-      DeleteIndexDefinition(index)
-
-    //Object
-    def objectId(objectId: String) =
-      DeleteObjectDefinition(oid = Some(objectId))
-
-    //Object
-    def from(index: String) = DeleteObjectDefinition(index = Some(index))
-
-    def key(keyName: String) = DeleteKeyDefinition(keyName)
-
-    def synonym(synId: String) = DeleteSynonymDefinition(synId = synId)
-
-    def rule(ruleId: String) = DeleteRuleDefinition(objectId = ruleId)
-
-  }
-
-  implicit object DeleteObjectDefinitionExecutable
-      extends Executable[DeleteObjectDefinition, Task] {
-    override def apply(client: AlgoliaClient, query: DeleteObjectDefinition)(
-        implicit executor: ExecutionContext): Future[Task] = {
-      client.request[Task](query.build())
-    }
-  }
-
-  implicit object DeleteIndexDefinitionExecutable extends Executable[DeleteIndexDefinition, Task] {
-    override def apply(client: AlgoliaClient, query: DeleteIndexDefinition)(
-        implicit executor: ExecutionContext): Future[Task] = {
-      client.request[Task](query.build())
-    }
-  }
-
-  implicit object DeleteByDefinitionExecutable extends Executable[DeleteByDefinition, Task] {
-    override def apply(client: AlgoliaClient, query: DeleteByDefinition)(
-        implicit executor: ExecutionContext): Future[Task] = {
-      client.request[Task](query.build())
-    }
-  }
-
-}
