+++ src/main/scala/algolia/definitions/GetDefinition.scala
@@ -90,50 +90,3 @@ case class GetObjectsDefinition(
   }
 
 }
-
-trait GetDsl {
-
-  implicit val formats: Formats
-
-  case object get {
-
-    def objectId(objectId: String) = GetObjectDefinition(oid = Some(objectId))
-
-    def from(index: String) = GetObjectDefinition(index = Some(index))
-
-    def key(keyName: String) = GetKeyDefinition(keyName)
-
-    @deprecated("use list keys", "1.27.1")
-    def allKeys() = ListKeysDefinition()
-
-    @deprecated("use list keysFrom", "1.27.1")
-    def allKeysFrom(indexName: String) =
-      ListKeysDefinition(indexName = Some(indexName))
-
-    def synonym(synId: String) = GetSynonymDefinition(synId = synId)
-
-    def rule(ruleId: String) = GetRuleDefinition(objectId = ruleId)
-
-  }
-
-  implicit object GetObjectDefinitionExecutable
-      extends Executable[GetObjectDefinition, GetObject] {
-
-    override def apply(client: AlgoliaClient, query: GetObjectDefinition)(
-        implicit executor: ExecutionContext): Future[GetObject] = {
-      client.request[JObject](query.build()).map(GetObject)
-    }
-
-  }
-
-  implicit object GetObjectsDefinitionExecutable
-      extends Executable[GetObjectsDefinition, Results] {
-
-    override def apply(client: AlgoliaClient, query: GetObjectsDefinition)(
-        implicit executor: ExecutionContext): Future[Results] = {
-      client.request[Results](query.build())
-    }
-
-  }
-
-}
