+++ src/main/scala/algolia/definitions/IndexSettingsDefinition.scala
@@ -91,44 +91,3 @@ case class IndexChangeSettingsDefinition(
     )
   }
 }
-
-trait IndexSettingsDsl {
-
-  implicit val formats: Formats
-
-  case object settings {
-
-    def of(index: String) = IndexSettingsDefinition(index)
-
-  }
-
-  @deprecated("use setSettings", "1.27.1")
-  case object changeSettings {
-
-    def of(index: String) = IndexSettingsDefinition(index)
-
-  }
-
-  case object setSettings {
-
-    def of(index: String) = IndexSettingsDefinition(index)
-
-  }
-
-  implicit object IndexSettingsDefinitionExecutable
-      extends Executable[IndexSettingsDefinition, IndexSettings] {
-    override def apply(client: AlgoliaClient, settings: IndexSettingsDefinition)(
-        implicit executor: ExecutionContext): Future[IndexSettings] = {
-      client.request[IndexSettings](settings.build())
-    }
-  }
-
-  implicit object IndexChangeSettingsDefinitionExecutable
-      extends Executable[IndexChangeSettingsDefinition, Task] {
-    override def apply(client: AlgoliaClient, settings: IndexChangeSettingsDefinition)(
-        implicit executor: ExecutionContext): Future[Task] = {
-      client.request[Task](settings.build())
-    }
-  }
-
-}
