+++ src/main/scala/algolia/definitions/KeyDefinition.scala
@@ -176,43 +176,3 @@ case class ListKeysDefinition(indexName: Option[String] = None,
     )
   }
 }
-
-trait KeyDefinitionDsl {
-
-  implicit val formats: Formats
-
-  implicit object GetApiKeyDefinitionExecutable extends Executable[GetKeyDefinition, ApiKey] {
-    override def apply(client: AlgoliaClient, query: GetKeyDefinition)(
-        implicit executor: ExecutionContext): Future[ApiKey] =
-      client.request[ApiKey](query.build())
-  }
-
-  implicit object AddApiKeyDefinitionExecutable
-      extends Executable[AddKeyDefinition, CreateUpdateKey] {
-    override def apply(client: AlgoliaClient, query: AddKeyDefinition)(
-        implicit executor: ExecutionContext): Future[CreateUpdateKey] =
-      client.request[CreateUpdateKey](query.build())
-  }
-
-  implicit object DeleteApiKeyDefinitionExecutable
-      extends Executable[DeleteKeyDefinition, DeleteKey] {
-    override def apply(client: AlgoliaClient, query: DeleteKeyDefinition)(
-        implicit executor: ExecutionContext): Future[DeleteKey] =
-      client.request[DeleteKey](query.build())
-  }
-
-  implicit object UpdateApiKeyDefinitionExecutable
-      extends Executable[UpdateKeyDefinition, CreateUpdateKey] {
-    override def apply(client: AlgoliaClient, query: UpdateKeyDefinition)(
-        implicit executor: ExecutionContext): Future[CreateUpdateKey] =
-      client.request[CreateUpdateKey](query.build())
-  }
-
-  implicit object GetAllApiKeyDefinitionExecutable
-      extends Executable[ListKeysDefinition, AllKeys] {
-    override def apply(client: AlgoliaClient, query: ListKeysDefinition)(
-        implicit executor: ExecutionContext): Future[AllKeys] =
-      client.request[AllKeys](query.build())
-  }
-
-}
