+++ src/main/scala/algolia/definitions/PartialUpdateObjectOperationDefinition.scala
@@ -180,87 +180,3 @@ case class PartialUpdateOneObjectDefinition(
     copy(`object` = Some(obj))
 
 }
-
-trait PartialUpdateObjectDsl {
-
-  implicit val formats: Formats
-
-  case object increment {
-
-    def attribute(attribute: String): PartialUpdateObjectOperationDefinition =
-      PartialUpdateObjectOperationDefinition(Increment, attribute = Some(attribute))
-
-  }
-
-  case object decrement {
-
-    def attribute(attribute: String): PartialUpdateObjectOperationDefinition =
-      PartialUpdateObjectOperationDefinition(Decrement, attribute = Some(attribute))
-
-  }
-
-  case object add {
-
-    def value(value: String): PartialUpdateObjectOperationDefinition =
-      PartialUpdateObjectOperationDefinition(Add, value = Some(value))
-
-    def key(key: ApiKey) = AddKeyDefinition(key)
-
-  }
-
-  case object addUnique {
-
-    def value(value: String): PartialUpdateObjectOperationDefinition =
-      PartialUpdateObjectOperationDefinition(AddUnique, value = Some(value))
-
-  }
-
-  case object remove {
-
-    def value(value: String): PartialUpdateObjectOperationDefinition =
-      PartialUpdateObjectOperationDefinition(Remove, value = Some(value))
-
-  }
-
-  case object update {
-
-    def attribute(attribute: String): PartialUpdateObjectDefinition =
-      PartialUpdateObjectDefinition(attribute = Some(attribute))
-
-    def key(keyName: String) = UpdateKeyDefinition(keyName)
-
-  }
-
-  case object partialUpdate {
-
-    def from(index: String): PartialUpdateOneObjectDefinition = {
-      PartialUpdateOneObjectDefinition(index = index)
-    }
-
-  }
-
-  implicit object PartialUpdateObjectOperationExecutable
-      extends Executable[PartialUpdateObjectOperationDefinition, Task] {
-    override def apply(client: AlgoliaClient, query: PartialUpdateObjectOperationDefinition)(
-        implicit executor: ExecutionContext): Future[Task] = {
-      client.request[Task](query.build())
-    }
-  }
-
-  implicit object PartialUpdateObjectExecutable
-      extends Executable[PartialUpdateObjectDefinition, Task] {
-    override def apply(client: AlgoliaClient, query: PartialUpdateObjectDefinition)(
-        implicit executor: ExecutionContext): Future[Task] = {
-      client.request[Task](query.build())
-    }
-  }
-
-  implicit object PartialUpdateOneObjectDefinitionExecutable
-      extends Executable[PartialUpdateOneObjectDefinition, Task] {
-    override def apply(client: AlgoliaClient, query: PartialUpdateOneObjectDefinition)(
-        implicit executor: ExecutionContext): Future[Task] = {
-      client.request[Task](query.build())
-    }
-  }
-
-}
