+++ src/main/scala/algolia/definitions/RulesDefinition.scala
@@ -231,64 +231,3 @@ case class SearchRulesDefinition(
     )
   }
 }
-
-trait RulesDsl {
-
-  implicit val formats: Formats
-
-  implicit object GetRuleDefinitionExecutable extends Executable[GetRuleDefinition, Rule] {
-
-    override def apply(client: AlgoliaClient, query: GetRuleDefinition)(
-        implicit executor: ExecutionContext): Future[Rule] = {
-      client.request[Rule](query.build())
-    }
-
-  }
-
-  implicit object DeleteRuleDefinitionExecutable extends Executable[DeleteRuleDefinition, Task] {
-
-    override def apply(client: AlgoliaClient, query: DeleteRuleDefinition)(
-        implicit executor: ExecutionContext): Future[Task] = {
-      client.request[Task](query.build())
-    }
-
-  }
-
-  implicit object ClearRulesDefinitionExecutable extends Executable[ClearRulesDefinition, Task] {
-
-    override def apply(client: AlgoliaClient, query: ClearRulesDefinition)(
-        implicit executor: ExecutionContext): Future[Task] = {
-      client.request[Task](query.build())
-    }
-
-  }
-
-  implicit object SaveRuleDefinitionExecutable extends Executable[SaveRuleDefinition, Task] {
-
-    override def apply(client: AlgoliaClient, query: SaveRuleDefinition)(
-        implicit executor: ExecutionContext): Future[Task] = {
-      client.request[Task](query.build())
-    }
-
-  }
-
-  implicit object BatchRulesDefinitionExecutable extends Executable[BatchRulesDefinition, Task] {
-
-    override def apply(client: AlgoliaClient, query: BatchRulesDefinition)(
-        implicit executor: ExecutionContext): Future[Task] = {
-      client.request[Task](query.build())
-    }
-
-  }
-
-  implicit object SearchRulesDefinitionExecutable
-      extends Executable[SearchRulesDefinition, SearchRuleResult] {
-
-    override def apply(client: AlgoliaClient, query: SearchRulesDefinition)(
-        implicit executor: ExecutionContext): Future[SearchRuleResult] = {
-      client.request[SearchRuleResult](query.build())
-    }
-
-  }
-
-}
