+++ src/main/scala/algolia/definitions/SynonymsDefinition.scala
@@ -231,84 +231,3 @@ case class SearchSynonymsDefinition(
     )
   }
 }
-
-trait SynonymsDsl {
-
-  implicit val formats: Formats
-
-  object save {
-
-    def synonym(synonym: AbstractSynonym) =
-      SaveSynonymDefinition(synonym = synonym)
-
-    def synonyms(synonyms: Iterable[AbstractSynonym]) =
-      BatchSynonymsDefinition(synonyms = synonyms)
-
-    def rule(rule: Rule) =
-      SaveRuleDefinition(rule = rule)
-
-    def rules(rules: Iterable[Rule]) =
-      BatchRulesDefinition(rules = rules)
-
-  }
-
-  implicit object GetSynonymDefinitionExecutable
-      extends Executable[GetSynonymDefinition, AbstractSynonym] {
-
-    override def apply(client: AlgoliaClient, query: GetSynonymDefinition)(
-        implicit executor: ExecutionContext): Future[AbstractSynonym] = {
-      client.request[AbstractSynonym](query.build())
-    }
-
-  }
-
-  implicit object DeleteSynonymDefinitionExecutable
-      extends Executable[DeleteSynonymDefinition, Task] {
-
-    override def apply(client: AlgoliaClient, query: DeleteSynonymDefinition)(
-        implicit executor: ExecutionContext): Future[Task] = {
-      client.request[Task](query.build())
-    }
-
-  }
-
-  implicit object ClearSynonymsDefinitionExecutable
-      extends Executable[ClearSynonymsDefinition, Task] {
-
-    override def apply(client: AlgoliaClient, query: ClearSynonymsDefinition)(
-        implicit executor: ExecutionContext): Future[Task] = {
-      client.request[Task](query.build())
-    }
-
-  }
-
-  implicit object SaveSynonymDefinitionExecutable extends Executable[SaveSynonymDefinition, Task] {
-
-    override def apply(client: AlgoliaClient, query: SaveSynonymDefinition)(
-        implicit executor: ExecutionContext): Future[Task] = {
-      client.request[Task](query.build())
-    }
-
-  }
-
-  implicit object BatchSynonymsDefinitionExecutable
-      extends Executable[BatchSynonymsDefinition, Task] {
-
-    override def apply(client: AlgoliaClient, query: BatchSynonymsDefinition)(
-        implicit executor: ExecutionContext): Future[Task] = {
-      client.request[Task](query.build())
-    }
-
-  }
-
-  implicit object SearchSynonymsDefinitionExecutable
-      extends Executable[SearchSynonymsDefinition, SearchSynonymResult] {
-
-    override def apply(client: AlgoliaClient, query: SearchSynonymsDefinition)(
-        implicit executor: ExecutionContext): Future[SearchSynonymResult] = {
-      client.request[SearchSynonymResult](query.build())
-    }
-
-  }
-
-}
