+++ src/main/scala/algolia/definitions/WaitForTaskDefinition.scala
@@ -57,75 +57,4 @@ case class WaitForTaskDefinition(taskId: Long,
 
 }
 
-trait WaitForTaskDsl {
-
-  case object waitFor {
-    def task(task: AlgoliaTask): WaitForTaskDefinition =
-      WaitForTaskDefinition(task.idToWaitFor())
-
-    def task(taskID: Long): WaitForTaskDefinition =
-      WaitForTaskDefinition(taskID)
-  }
-
-  implicit object WaitForTaskDefinitionExecutable
-      extends Executable[WaitForTaskDefinition, TaskStatus] {
-
-    // Run every 100 ms, use a wheel with 512 buckets
-    private lazy val timer = {
-      val threadFactory = new ThreadFactory {
-        override def newThread(r: Runnable): Thread = {
-          val t = Executors.defaultThreadFactory().newThread(r)
-          t.setDaemon(true)
-          t.setName("algolia-waitfor-thread")
-          t
-        }
-      }
-      new HashedWheelTimer(threadFactory, 100, TimeUnit.MILLISECONDS, 512)
-    }
-
-    /**
-      * Wait for the completion of a task
-      * By default it waits `baseDelay` and multiple it by 2 each time until it's greater than `maxDelay`:
-      *  100     =  100ms
-      *  100 * 2 =  200ms
-      *  200 * 2 =  400ms
-      *  400 * 2 =  800ms
-      *  800 * 2 = 1600ms
-      * 1600 * 2 = 3200ms
-      * 3200 * 2 = 6400ms
-      * etc...
-      *
-      */
-    override def apply(client: AlgoliaClient, query: WaitForTaskDefinition)(
-        implicit executor: ExecutionContext): Future[TaskStatus] = {
-
-      def request(d: Long, totalDelay: Long): Future[TaskStatus] =
-        delay[TaskStatus](d) {
-          client.request[TaskStatus](query.build())
-        }.flatMap { res =>
-          if (res.status == "published") {
-            Future.successful(res)
-          } else if (totalDelay > query.maxDelay) {
-            Future.failed(WaitForTimeoutException(
-              s"Waiting for task `${query.taskId}` on index `${query.index.get}` timeout after ${d}ms"))
-          } else {
-            request(d * 2, totalDelay + d)
-          }
-        }
-
-      request(query.baseDelay, 0L)
-    }
-
-    private def delay[T](delay: Long)(block: => Future[T]): Future[T] = {
-      val promise = Promise[T]()
-      val task = new TimerTask {
-        override def run(timeout: Timeout): Unit = promise.completeWith(block)
-      }
-      timer.newTimeout(task, delay, TimeUnit.MILLISECONDS)
-      promise.future
-    }
-  }
-
-}
-
 case class WaitForTimeoutException(message: String) extends Exception(message)
