+++ src/main/scala/algolia/dsl/IndexSettingsDsl.scala
@@ -23,41 +23,53 @@
  * THE SOFTWARE.
  */
 
-package algolia.definitions
+package algolia.dsl
 
-import algolia.http.{GET, HttpPayload}
-import algolia.objects.{Query, RequestOptions}
-import algolia.responses.BrowseResult
+import algolia.definitions.{IndexChangeSettingsDefinition, IndexSettingsDefinition}
+import algolia.objects.IndexSettings
+import algolia.responses.Task
 import algolia.{AlgoliaClient, Executable}
 import org.json4s.Formats
 
 import scala.concurrent.{ExecutionContext, Future}
 
-case class BrowseIndexDefinition(
-    source: String,
-    query: Option[Query] = None,
-    cursor: Option[String] = None,
-    requestOptions: Option[RequestOptions] = None)(implicit val formats: Formats)
-    extends Definition {
+trait IndexSettingsDsl {
 
-  type T = BrowseIndexDefinition
+  implicit val formats: Formats
 
-  def from(cursor: String): BrowseIndexDefinition = copy(cursor = Some(cursor))
+  case object settings {
 
-  def query(query: Query): BrowseIndexDefinition = copy(query = Some(query))
+    def of(index: String) = IndexSettingsDefinition(index)
 
-  override def options(requestOptions: RequestOptions): BrowseIndexDefinition =
-    copy(requestOptions = Some(requestOptions))
-
-  override private[algolia] def build(): HttpPayload = {
-    val q = query.getOrElse(Query()).copy(cursor = cursor)
-
-    HttpPayload(
-      GET,
-      Seq("1", "indexes", source, "browse"),
-      queryParameters = Some(q.toQueryParam),
-      isSearch = true,
-      requestOptions = requestOptions
-    )
   }
+
+  @deprecated("use setSettings", "1.27.1")
+  case object changeSettings {
+
+    def of(index: String) = IndexSettingsDefinition(index)
+
+  }
+
+  case object setSettings {
+
+    def of(index: String) = IndexSettingsDefinition(index)
+
+  }
+
+  implicit object IndexSettingsDefinitionExecutable
+      extends Executable[IndexSettingsDefinition, IndexSettings] {
+    override def apply(client: AlgoliaClient, settings: IndexSettingsDefinition)(
+        implicit executor: ExecutionContext): Future[IndexSettings] = {
+      client.request[IndexSettings](settings.build())
+    }
+  }
+
+  implicit object IndexChangeSettingsDefinitionExecutable
+      extends Executable[IndexChangeSettingsDefinition, Task] {
+    override def apply(client: AlgoliaClient, settings: IndexChangeSettingsDefinition)(
+        implicit executor: ExecutionContext): Future[Task] = {
+      client.request[Task](settings.build())
+    }
+  }
+
 }
