+++ src/main/scala/algolia/dsl/IndexingBatchDsl.scala
@@ -23,48 +23,22 @@
  * THE SOFTWARE.
  */
 
-package algolia.definitions
+package algolia.dsl
 
-import algolia.http.{HttpPayload, POST}
-import algolia.inputs.{AddObjectOperation, BatchOperations, UpdateObjectOperation}
-import algolia.objects.RequestOptions
+import algolia.definitions.IndexingBatchDefinition
 import algolia.responses.TasksSingleIndex
 import algolia.{AlgoliaClient, Executable}
-import org.json4s.Formats
-import org.json4s.native.Serialization._
 
 import scala.concurrent.{ExecutionContext, Future}
 
-case class IndexingBatchDefinition(
-    index: String,
-    definitions: Traversable[Definition] = Traversable(),
-    requestOptions: Option[RequestOptions] = None)(implicit val formats: Formats)
-    extends Definition
-    with BatchOperationUtils {
+trait IndexingBatchDsl {
 
-  type T = IndexingBatchDefinition
-
-  override def options(requestOptions: RequestOptions): IndexingBatchDefinition =
-    copy(requestOptions = Some(requestOptions))
-
-  override private[algolia] def build(): HttpPayload = {
-    val operations = definitions.map {
-      case IndexingDefinition(_, None, Some(obj), _) =>
-        hasObjectId(obj) match {
-          case (true, o) => UpdateObjectOperation(o)
-          case (false, o) => AddObjectOperation(o)
-        }
-
-      case IndexingDefinition(_, Some(objectId), Some(obj), _) =>
-        UpdateObjectOperation(addObjectId(obj, objectId))
+  implicit object IndexingBatchDefinitionExecutable
+      extends Executable[IndexingBatchDefinition, TasksSingleIndex] {
+    override def apply(client: AlgoliaClient, batch: IndexingBatchDefinition)(
+        implicit executor: ExecutionContext): Future[TasksSingleIndex] = {
+      client.request[TasksSingleIndex](batch.build())
     }
-
-    HttpPayload(
-      POST,
-      Seq("1", "indexes", index, "batch"),
-      body = Some(write(BatchOperations(operations))),
-      isSearch = false,
-      requestOptions = requestOptions
-    )
   }
+
 }
