+++ src/main/scala/algolia/dsl/MoveDsl.scala
@@ -23,41 +23,30 @@
  * THE SOFTWARE.
  */
 
-package algolia.definitions
+package algolia.dsl
 
-import algolia.http.{GET, HttpPayload}
-import algolia.objects.{Query, RequestOptions}
-import algolia.responses.BrowseResult
+import algolia.definitions.MoveIndexDefinition
+import algolia.responses.Task
 import algolia.{AlgoliaClient, Executable}
 import org.json4s.Formats
 
 import scala.concurrent.{ExecutionContext, Future}
 
-case class BrowseIndexDefinition(
-    source: String,
-    query: Option[Query] = None,
-    cursor: Option[String] = None,
-    requestOptions: Option[RequestOptions] = None)(implicit val formats: Formats)
-    extends Definition {
+trait MoveDsl {
 
-  type T = BrowseIndexDefinition
+  implicit val formats: Formats
 
-  def from(cursor: String): BrowseIndexDefinition = copy(cursor = Some(cursor))
+  case object move {
 
-  def query(query: Query): BrowseIndexDefinition = copy(query = Some(query))
+    def index(index: String): MoveIndexDefinition = MoveIndexDefinition(index)
 
-  override def options(requestOptions: RequestOptions): BrowseIndexDefinition =
-    copy(requestOptions = Some(requestOptions))
-
-  override private[algolia] def build(): HttpPayload = {
-    val q = query.getOrElse(Query()).copy(cursor = cursor)
-
-    HttpPayload(
-      GET,
-      Seq("1", "indexes", source, "browse"),
-      queryParameters = Some(q.toQueryParam),
-      isSearch = true,
-      requestOptions = requestOptions
-    )
   }
+
+  implicit object MoveIndexDefinitionExecutable extends Executable[MoveIndexDefinition, Task] {
+    override def apply(client: AlgoliaClient, query: MoveIndexDefinition)(
+        implicit executor: ExecutionContext): Future[Task] = {
+      client.request[Task](query.build())
+    }
+  }
+
 }
