+++ src/main/scala/algolia/dsl/SearchDsl.scala
@@ -23,41 +23,48 @@
  * THE SOFTWARE.
  */
 
-package algolia.definitions
+package algolia.dsl
 
-import algolia.http.{GET, HttpPayload}
-import algolia.objects.{Query, RequestOptions}
-import algolia.responses.BrowseResult
+import algolia.AlgoliaDsl.In
+import algolia.definitions.{
+  SearchDefinition,
+  SearchFacetDefinition,
+  SearchRulesDefinition,
+  SearchSynonymsDefinition
+}
+import algolia.responses.{SearchFacetResult, SearchResult}
 import algolia.{AlgoliaClient, Executable}
 import org.json4s.Formats
 
 import scala.concurrent.{ExecutionContext, Future}
 
-case class BrowseIndexDefinition(
-    source: String,
-    query: Option[Query] = None,
-    cursor: Option[String] = None,
-    requestOptions: Option[RequestOptions] = None)(implicit val formats: Formats)
-    extends Definition {
+trait SearchDsl {
 
-  type T = BrowseIndexDefinition
+  implicit val formats: Formats
 
-  def from(cursor: String): BrowseIndexDefinition = copy(cursor = Some(cursor))
+  case object search {
 
-  def query(query: Query): BrowseIndexDefinition = copy(query = Some(query))
+    def into(index: String) = SearchDefinition(index)
 
-  override def options(requestOptions: RequestOptions): BrowseIndexDefinition =
-    copy(requestOptions = Some(requestOptions))
+    def synonyms(i: In): SearchSynonymsDefinition = SearchSynonymsDefinition()
 
-  override private[algolia] def build(): HttpPayload = {
-    val q = query.getOrElse(Query()).copy(cursor = cursor)
+    def rules(i: In): SearchRulesDefinition = SearchRulesDefinition()
 
-    HttpPayload(
-      GET,
-      Seq("1", "indexes", source, "browse"),
-      queryParameters = Some(q.toQueryParam),
-      isSearch = true,
-      requestOptions = requestOptions
-    )
   }
+
+  implicit object SearchDefinitionExecutable extends Executable[SearchDefinition, SearchResult] {
+    override def apply(client: AlgoliaClient, query: SearchDefinition)(
+        implicit executor: ExecutionContext): Future[SearchResult] = {
+      client.request[SearchResult](query.build())
+    }
+  }
+
+  implicit object SearchFacetDefinitionExecutable
+      extends Executable[SearchFacetDefinition, SearchFacetResult] {
+    override def apply(client: AlgoliaClient, query: SearchFacetDefinition)(
+        implicit executor: ExecutionContext): Future[SearchFacetResult] = {
+      client.request[SearchFacetResult](query.build())
+    }
+  }
+
 }
